Move expansion result types to yash-core
